refactor(shell): export missing primitives from the barrel
