+++ src/components/shell/index.ts
@@ -1,6 +1,10 @@
 export { default as AppShell } from './AppShell.svelte';
+export { default as AsideToggleButton } from './AsideToggleButton.svelte';
+export { default as ChannelRow } from './ChannelRow.svelte';
 export { default as ChannelSidebar } from './ChannelSidebar.svelte';
 export { default as ChatFrame } from './ChatFrame.svelte';
 export { default as MemberSidebar } from './MemberSidebar.svelte';
 export { default as NavRail } from './NavRail.svelte';
+export { default as PresenceAvatar } from './PresenceAvatar.svelte';
+export { default as RailTile } from './RailTile.svelte';
 export { default as ShellDrawer } from './ShellDrawer.svelte';
